+++ src/Monofony/Pack/AdminPack/.recipe/tests/Behat/Page/Backend/Administrator/CreatePage.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace App\Tests\Behat\Page\Backend\Administrator;
 
-use Monofony\Component\Admin\Behat\Crud\AbstractCreatePage;
-use Monofony\Component\Admin\Behat\Crud\CreatePageInterface;
+use Monofony\Bridge\Behat\Crud\AbstractCreatePage;
+use Monofony\Bridge\Behat\Crud\CreatePageInterface;
 
 class CreatePage extends AbstractCreatePage implements CreatePageInterface
 {
