+++ src/Monofony/Pack/AdminPack/.recipe/tests/Behat/Page/Backend/Customer/UpdatePage.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\Tests\Behat\Page\Backend\Customer;
 
 use Behat\Mink\Element\NodeElement;
-use Monofony\Component\Admin\Behat\Crud\AbstractUpdatePage;
-use Monofony\Component\Admin\Behat\Crud\UpdatePageInterface;
+use Monofony\Bridge\Behat\Crud\AbstractUpdatePage;
+use Monofony\Bridge\Behat\Crud\UpdatePageInterface;
 use Monofony\Bridge\Behat\Behaviour\Toggles;
 
 class UpdatePage extends AbstractUpdatePage implements UpdatePageInterface
